Cast Plex rating to float64
Convert anime.Plex.Rating to float64 before passing it to UpdateAnimeScore to match the expected parameter type and avoid a type mismatch. Preserves the existing nil-check and default rating of 0.0.

+++ internal/animeupdate/service.go
@@ -191,7 +191,7 @@ func (s *service) syncToAniList(ctx context.Context, anime *domain.AnimeUpdate, 
 	} else {
 		rating := 0.0
 		if anime.Plex != nil {
-			rating = anime.Plex.Rating
+			rating = float64(anime.Plex.Rating)
 		}
 		if err := s.anilistAuthService.UpdateAnimeScore(ctx, anilistID, rating); err != nil {
 			s.log.Warn().Err(err).Int("anilistID", anilistID).Msg("AniList: failed to update score")
